// src/_old/types.ts
export interface TowerUpgrade {
  acceleration?: number;
  aim?: number;
  all_seeing?: number;
  armor_destruction?: number;
  armor_piercing?: boolean;
  beams?: number;
  bleed_increase?: number;
  bleed?: number;
  bliz_length?: number;
  bliz_radius?: number;
  boss_slayer?: number;
  bounty_add?: number;
  bounty_amp?: number;
  burn_damage?: number;
  burn_time?: number;
  burst_cd?: number;
  burst?: number;
  camo?: boolean;
  cash_conv?: number;
  cooldown_buff?: number;
  cooldown?: number;
  corrosion?: number;
  crit_chance?: number;
  crit_dmg?: number;
  damage_buff?: number;
  damage_cap?: number;
  damage_increase?: number;
  damage?: number;
  departure?: number;
  exclusion?: number;
  execution?: number;
  first_strike?: number;
  frost_time?: number;
  frostbite?: number;
  hollow_point?: number;
  inaccuracy?: number;
  income?: number;
  instabreak?: number;
  interest?: number;
  intimidation?: number;
  leech?: number;
  max_cooldown?: number;
  max_dmg_increase?: number;
  max_interest?: number;
  max_spawns?: number;
  melt?: number;
  min_cooldown?: number;
  min_dmg_increase?: number;
  oblivion?: number;
  paralyze?: number;
  price?: number;
  purify?: number;
  rad_length?: number;
  rad_strength?: number;
  radius?: number;
  range_buff?: number;
  range?: number;
  residue_length?: number;
  rupture_duration?: number;
  rupture?: number;
  scald?: number;
  scorch_length?: number;
  scorch?: number;
  shield_bypass?: number;
  shield_dmg?: number;
  shredding_length?: number;
  shredding?: number;
  soul_damage?: number;
  spawns_cd?: number;
  spawns_damage?: number;
  spawns_range?: number;
  stun?: number;
  targeting_rate?: number;
  taxxing?: number;
  upgrade_reduction?: number;
  valor?: number;
}